+++ src/main/java/com/kindergarten/api/config/WebSecurityConfig.java
@@ -46,6 +46,7 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/api/users/**").permitAll()
                 .antMatchers("/api/users/existid/**").permitAll()
                 .antMatchers("/api/kindergartens/**").permitAll()
+                .antMatchers("/api/api/signup-kindergartens/**").permitAll()
                 .antMatchers("/api/auth/**").permitAll()
                 .antMatchers("/api/studnet/**").permitAll()
                 .antMatchers(HttpMethod.GET, "/exception/**", "/actuator/health", "/favicon.ico").permitAll()
